# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/grader-lab:
    post:
      operationId: runGraderLab
      tags: [admin]
      summary: Admin debug — generate a spread of candidate answers for a problem and grade each (no AI detection)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GraderLabInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GraderLabResult"
...
  /admin/settings:
    put:
      operationId: updateCourseSettings
      tags: [admin]
      summary: Update operator-tunable course settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSettingsInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CourseSettings"
...
components:
  schemas:
...
    GraderLabInput:
      type: object
      description: Provide a problemId to audit a real problem, or a custom prompt + correctAnswer.
      properties:
        problemId: { type: ["integer", "null"] }
        prompt: { type: ["string", "null"] }
        correctAnswer: { type: ["string", "null"] }
    GraderLabResult:
      type: object
      required: [prompt, correctAnswer, cases]
      properties:
        prompt: { type: string }
        correctAnswer: { type: string }
        cases:
          type: array
          items:
            $ref: "#/components/schemas/GraderLabCase"
    GraderLabCase:
      type: object
      required: [label, kind, expectedCredit, answer, credit, explanation]
      properties:
        label: { type: string }
        kind: { type: string }
        # Expected partial credit in [0, 1] under the inverted CCR scale.
        expectedCredit: { type: number }
        answer: { type: string }
        # Credit the grader actually awarded in [0, 1].
        credit: { type: number }
        explanation: { type: string }
    CourseSettings:
      type: object
      required: [formatWeightMcq, formatWeightHybrid, formatWeightWritten, minDiagnostics]
      properties:
        formatWeightMcq: { type: number }
        formatWeightHybrid: { type: number }
        formatWeightWritten: { type: number }
        minDiagnostics: { type: integer }
    UpdateSettingsInput:
      type: object
      properties:
        formatWeightMcq: { type: ["number", "null"] }
        formatWeightHybrid: { type: ["number", "null"] }
        formatWeightWritten: { type: ["number", "null"] }
        minDiagnostics: { type: ["integer", "null"] }
...